Switch back to XMPP from matrix
After ~4 years of on-off trial matrix seems to still not be usable, if
ever. Shelving it for now, maybe will see where it is in a few more
years, or if the protocol design is flawed.
Latency is often unusable, rooms are frequently getting corrupted state,
clients lock up / crash all the time and have memory usage issues. Along
with seeming to not really provide many actually useful moderation /
community features. Many bugs are years old and more half working
features are getting tacked on without need or fixing the issues with
the existing ones.
The servers have just as many issues and running one is unreasonably
difficult. Bridging also has it's own problems.

At least XMPP sort of works with a very select feature set and a few
hours of setup, especially now with OMEMO, even if the protocol/spec is
a disaster.